Minor Improvements
Add Witch Alias to Mage
Make damage a double

+++ src/com/gmail/scyntrus/tmob/mobs/Archer.java
@@ -54,7 +54,7 @@ public class Archer extends EntitySkeleton implements TownyMob {
 	public static double powerCost = 0;
 	public static double moneyCost = 0;
 	public static double range = 16;
-	public static int damage = 0;
+	public static double damage = 0;
 	public static int drops = 0;
 	private int retargetTime = 0;
 	private double moveSpeed;
@@ -455,7 +455,7 @@ public class Archer extends EntitySkeleton implements TownyMob {
 	@Override
 	public void a(EntityLiving entityliving, float f) { //TODO: Update name on version change
 		if (damage>0) {
-			super.a(entityliving, damage/2F);
+			super.a(entityliving, (float) damage/2F);
 		} else {
 			super.a(entityliving, f);
 		}
